+++ Example/WalletApp/PresentationLayer/Wallet/AuthRequest/AuthRequestPresenter.swift
@@ -53,7 +53,7 @@ final class AuthRequestPresenter: ObservableObject {
     }
 
     @MainActor
-    func approve() async {
+    func signMulti() async {
         do {
             ActivityIndicatorManager.shared.start()
 
@@ -61,14 +61,13 @@ final class AuthRequestPresenter: ObservableObject {
 
             _ = try await Web3Wallet.instance.approveSessionAuthenticate(requestId: request.id, auths: auths)
             ActivityIndicatorManager.shared.stop()
-            router.dismiss()
             /* Redirect */
-//            if let uri = request.requester.redirect?.native {
-//                WalletConnectRouter.goBack(uri: uri)
-//                router.dismiss()
-//            } else {
-//                showSignedSheet.toggle()
-//            }
+            if let uri = request.requester.redirect?.native {
+                WalletConnectRouter.goBack(uri: uri)
+                router.dismiss()
+            } else {
+                showSignedSheet.toggle()
+            }
 
         } catch {
             ActivityIndicatorManager.shared.stop()
